Start ditching port/adapters by name
