+++ crates/cli/src/main.rs
@@ -3,9 +3,8 @@ mod config;
 
 use clap::Parser;
 use selfie::{
-    adapters::config_loader,
-    config::ports::{ApplyToConfg, ConfigLoader},
-    filesystem::adapter::RealFileSystem,
+    config::loader::{self, ApplyToConfg, ConfigLoader},
+    filesystem::real::RealFileSystem,
 };
 use tracing::debug;
 
@@ -21,7 +20,7 @@ fn main() -> anyhow::Result<()> {
 
     let config = {
         // 1. Load config.yaml
-        let config = config_loader::Yaml::new(&fs).load_config()?;
+        let config = loader::Yaml::new(&fs).load_config()?;
 
         // 2. Apply CLI args to config (overriding)
         args.apply_to_config(config)
