+++ crates/selfie/src/adapters.rs
@@ -1,2 +1,1 @@
 pub mod command_runner;
-pub mod config_loader;
