--- crates/selfie/src/adapters/config_loader.rs
@@ -1,3 +1,0 @@
-pub mod yaml;
-
-pub use self::yaml::Yaml;
