+++ crates/selfie/src/config.rs
@@ -1,4 +1,4 @@
-pub mod ports;
+pub mod loader;
 
 use std::{
     num::{NonZeroU64, NonZeroUsize},
